fix: customer api and publish customer action

+++ src/modules/customer/service/CustomerService.ts
@@ -2,20 +2,28 @@ import {inject, injectable} from "inversify";
 import {ICustomerService} from "./ICustomerService";
 import {ICustomerRepository} from "../repo/ICustomerRepository";
 import {TYPES} from "../../../types";
-import {createInternalError, Err} from "../../../libs/errors";
+import {createForbiddenError, createInternalError, Err} from "../../../libs/errors";
 import {err, ok, ResultAsync} from "neverthrow";
 import {CustomerCreate, customerCreateSchema} from "../entity/customerCreate";
 import {Validator} from "../../../libs/validator";
 import {IRequester} from "../../../libs/IRequester";
 import {CustomerUpdate, customerUpdateSchema} from "../entity/customerUpdate";
 import {Customer} from "../entity/customer";
+import {CustomerFilter} from "../entity/customerFilter";
+import {Paging} from "../../../libs/paging";
+import {createMessage, IPubSub} from "../../../components/pubsub";
+import {topicCreateCustomer, topicDeleteCustomer, topicUpdateCustomer} from "../../../libs/topics";
+import {SystemRole} from "../../user/entity/user";
 
 @injectable()
 export class CustomerService implements ICustomerService{
-    constructor(@inject(TYPES.ICustomerRepository) private readonly customerRepository: ICustomerRepository) {
+    constructor(
+        @inject(TYPES.ICustomerRepository) private readonly customerRepository: ICustomerRepository,
+        @inject(TYPES.IPubSub) private readonly pubSub: IPubSub
+    ) {
     }
 
-    create = (requester: IRequester, c: CustomerCreate): ResultAsync<Customer, Err> => {
+    create(requester: IRequester, c: CustomerCreate): ResultAsync<Customer, Err> {
         return ResultAsync.fromPromise(
             (async () => {
                 const vR = (await Validator(customerCreateSchema, c))
@@ -28,12 +36,14 @@ export class CustomerService implements ICustomerService{
                     return err(r.error);
                 }
 
+                this.pubSub.Publish(topicCreateCustomer, createMessage(c, requester));
+
                 return ok(r.value);
             })(), e => createInternalError(e)
         ).andThen(r => r)
     }
 
-    update = (requester: IRequester, id: string, c: CustomerUpdate): ResultAsync<Customer, Err> => {
+    update(requester: IRequester, id: number, c: CustomerUpdate): ResultAsync<Customer, Err> {
         return ResultAsync.fromPromise(
             (async () => {
                 const vR = (await Validator(customerUpdateSchema, c))
@@ -41,30 +51,48 @@ export class CustomerService implements ICustomerService{
                     return err(vR.error);
                 }
 
+                const old = await this.customerRepository.findById(id);
+                if (old.isErr()) {
+                    return err(old.error);
+                }
+
                 const r = await this.customerRepository.update(id, c);
                 if (r.isErr()) {
                     return err(r.error);
                 }
+
+                this.pubSub.Publish(topicUpdateCustomer, createMessage({
+                    id: id,
+                    old: old.value,
+                    new: c
+                }, requester));
 
                 return ok(r.value);
             })(), e => createInternalError(e)
         ).andThen(r => r)
     }
 
-    delete(requester: IRequester, id: string): ResultAsync<void, Err> {
+    delete(requester: IRequester, id: number): ResultAsync<void, Err> {
         return ResultAsync.fromPromise(
             (async () => {
+                const old = await this.customerRepository.findById(id);
+                if (old.isErr()) {
+                    return err(old.error);
+                }
+
                 const r = await this.customerRepository.delete(id);
                 if (r.isErr()) {
                     return err(r.error);
                 }
+
+                this.pubSub.Publish(topicDeleteCustomer, createMessage(old.value, requester))
 
                 return ok(undefined);
             })(), e => createInternalError(e)
         ).andThen(r => r)
     }
 
-    findById(requester: IRequester, id: string): ResultAsync<Customer, Err> {
+    findById(id: number): ResultAsync<Customer, Err> {
         return ResultAsync.fromPromise(
             (async () => {
                 const r = await this.customerRepository.findById(id);
@@ -77,10 +105,10 @@ export class CustomerService implements ICustomerService{
         ).andThen(r => r)
     }
 
-    list(requester: IRequester): ResultAsync<Customer[], Err> {
+    list(filter: CustomerFilter, page: Paging): ResultAsync<Customer[], Err> {
         return ResultAsync.fromPromise(
             (async () => {
-                const r = await this.customerRepository.list();
+                const r = await this.customerRepository.list(filter, page);
                 if (r.isErr()) {
                     return err(r.error);
                 }
@@ -90,7 +118,7 @@ export class CustomerService implements ICustomerService{
         ).andThen(r => r)
     }
 
-    increasePaymentAmount(id: string): ResultAsync<void, Err> {
+    increasePaymentAmount(id: number): ResultAsync<void, Err> {
         return ResultAsync.fromPromise(
             (async () => {
                 const r = await this.customerRepository.increasePaymentAmount(id);
